Simplify the "view as" user credentials
